api-server/requirements.txt
# Flask Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
Flask-Migrate>=4.0.0

# Database
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.1.0
python-dotenv>=1.0.0

# Validation
marshmallow>=3.20.0
marshmallow-sqlalchemy>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# Utilities
redis>=5.0.0
celery>=5.3.0
gunicorn>=21.2.0

# Logging & Monitoring
python-json-logger>=2.0.0
prometheus-client>=0.19.0
